requirements.txt
pytest==8.2.0
Appium-Python-Client==4.0.0
selene==2.0.0rc6
allure-pytest==2.13.5
python-dotenv==1.0.1
requests==2.31.0